-Backgrounds More work on equipment and inventory
- Equipment drops now working
- Equipment mostly working in inventory
- Backgrounds working again
- Small work on movement polish
- Muzzle flash now on bullet start frame
- Better Swarm explosion SFX
- Added some equipment types

+++ AsteroidAnnihilation/Assets/_Scripts/Enemies/Bosses/BossFire.cs
@@ -12,7 +12,7 @@ namespace AsteroidAnnihilation
         public float FireRate;
         public int Amount;
         public float Spread;
-        public int BulletSpeed;
+        public float BulletSpeed;
         public float ProjectileLifeTime = 1f;
         public EnumCollections.EnemyProjectiles Projectile;
         private ObjectPooler objectPooler;
